Add first changes for TF Protocol V2.0

+++ com/com.c
@@ -1,5 +1,5 @@
 /* bricklib
- * Copyright (C) 2009-2010 Olaf Lüke <olaf@tinkerforge.com>
+ * Copyright (C) 2009-2012 Olaf Lüke <olaf@tinkerforge.com>
  *
  * com.c: Definition of communication interfaces
  *
@@ -36,7 +36,7 @@ uint8_t com_last_ext_id[2] = {1, 1};
 ComType com_current = COM_NONE;
 ComType com_ext[2] = {COM_NONE, COM_NONE};
 uint8_t com_stack_id = 1;
-uint64_t com_brick_uid = 62;
+uint32_t com_brick_uid = 62;
 
 Com com_list[] = {
 	{COM_NONE, no_init, no_send, no_recv},
