+++ com/i2c/i2c_eeprom/i2c_eeprom_master.h
@@ -1,5 +1,5 @@
 /* bricklib
- * Copyright (C) 2010 Olaf Lüke <olaf@tinkerforge.com>
+ * Copyright (C) 2010-2012 Olaf Lüke <olaf@tinkerforge.com>
  *
  * i2c_eeprom_master.h: i2c eeprom master functionality (for M24C64)
  *
@@ -42,7 +42,7 @@ bool i2c_eeprom_master_write(Twi *twi,
                              const uint16_t length);
 
 void i2c_eeprom_master_init(Twi *twi);
-uint64_t i2c_eeprom_master_read_uid(Twi *twi);
+uint32_t i2c_eeprom_master_read_uid(Twi *twi);
 bool i2c_eeprom_master_write_uid(Twi *twi, uint64_t uid);
 
 bool i2c_eeprom_master_read_plugin(Twi *twi,
